Add the ability to change mark line width and hide help text
When marking elements that are close to each other in a photo the 2 pixel line width of a mark might cover the other element.
By allowing the user to change the width of the line between 1 to 3 pixel the user can now select the line width that works for them.
The help line became very long, I've also add the ability to toggle if the help is show or not.